Make possible for agent to read utp results
Agent can run connected tests and we need to provide results back.
At the moment we do not know the devices in advance so
we need to change TaskOutputProcessor slightly to be able
to listen for the results for any device id.
Provide back test cases results in minimal data structure
to avoid depending on utp protos from agent module direcly.

Bug: 441471884
Test: N/A, integration tests on gradle_build tool changes
Change-Id: I16b4b194cf42b7b88c350c8f8f411b9f89e74c99

GitOrigin-RevId: bb1c36c40dd08afc91dfcf5dbf166af042a417ba