Address misleading message in bundle generation
There are several reasons why it might not be possible to generate a
Bundle for the user's selection.  Now that we don't support AGP 3.2
projects, blaming it on too low a version of AGP is not reasonable.

Also, clean up the truly awful way the warning message was being
generated to show in the UI.

Bug: 216744700
Test: manual
Change-Id: I4b4c26f8d42478c0bc7a47a0abb1164e981694ea

GitOrigin-RevId: 8064b114a2b9545b2f44ca64e2ce7a255bca903a