[LE] Add logging to LiveUpdateErrors
We used to log LiveUpdateErrors in more places than we'd like.

There was an effort to clean up those logging messages to avoid
log polution after we finished all the UI to properly display
the errors to the user.

However, we should still log it at least ONCE for bug report
and test logging purposes.

Fixes: 440602487
Test: None. This is a follow up CL for a test fix.
Change-Id: I099a5189490d0e4784b1b4afb84c26b40681da6a

GitOrigin-RevId: 9eb10a2a0f009d8cd0c6b383817356c15d7179d7