temporarily include a custom rule in always_build_rule
_java_spanner_proto_library_rule generated jar but we will not include
it when it's in in-project target. To resolve it, we can name it in
always build rules. Even though the ideal solution should be always
build any unknown rules, it's not the topic of this cl. So temporarily
include it until we get a better solution soon.

Bug: b/422487918
Test: manually test, I will update a test project in separate cl
Change-Id: Ie3b7ccec1addd2f8d0d1dee3db7eb4802f918e0d

GitOrigin-RevId: ce499e2ea78cbe15394928861317403f378906c7
(cherry picked from commit 4b92b4a2b0dfbdca06035ccad634cb96fe12b3fe)

Space-RevId: e87a0894543731cfbdda335f26ca6d5fb68d80ab