Update memoryAxis range in Live Telementry
This CL adds the MemoryAxis to the list of updatable components in
LiveMemoryFootprintModel to ensure it is refreshed along with the memory usage data.

This change fixes a bug where the Y-axis of the memory graph in the
"View Live Telemetry" profiler view was not updating, leading to an incorrect scale.

Bug: 377230194
Test: N/A
Change-Id: Iedc3a7eacb59fc04445502dea20f9e6a74fbd3ff

GitOrigin-RevId: 1c91f931b110f133dafbdbbf6df91f4ae1ce4e8d
(cherry picked from commit b1d7acc2627d7bc7bc12b22795edb82494834869)

Space-RevId: b9fff15eb9dd652545bd8ad38eb682decd20aa3d