More restrictive @RunsInEdt for withIntegrationTestEnvironment
Running the whole of a test method, including project open, in the
Application's Event Dispatch Thread is both more than most tests
require, and likely to lead to problems on further changes to the
underlying threading model beyond the deadlocks that we have already
seen.  The common case for use of @RunsInEdt is so that some of the
actual test body can run without requiring insertion of explicit
`runInEdtAndWait {...}` or similar thread handling.

For withIntegrationTestEnvironment, then, we will achieve the desired
behaviour by affecting just the behaviour of

  openPreparedProject { p -> ... }

and specifically which thread the action { p -> ... } should be run
on, based only on whether the test is annotated with @RunsInEdt or
not.  (We want this because there are hundreds of tests written with
@RunsInEdt, while the EdtRule interpretation of @RunsInEdt is
incompatible with the way we prepare and open the project.)

In a language which supported dynamic binding, this would look
something like

  (defmacro define-maybe-edt-test ((name &key edt) &body body)
    `(define-test name
       (let ((*run-action-on-edt* ',edt))
         ,@body)))

along with

  (apply (if *run-action-on-edt* #'funcall-on-edt #'funcall) args)

at the point where the action is called.

We don't have a binding stack.  However, what we do have is an
IntegrationTestEnvironment, which is propagated essentially all
the way from the rule instantiation to the project open.  And, since
it's an Environment I don't feel too bad about stuffing something that
looks like a binding in there, for lookup at project open time.

So, extend IntegrationTestEnvironment with this property.  Arrange for
withIntegrationTestEnvironment() to populate it based on @RunsInEdt
annotations, and for project openers to respect the value.  And hey
presto, Bob is probably your uncle.

Adjust the relatively few tests that need adjusting for this new
interpretation of @RunsInEdt.

Bug: 430533136
Test: existing
Change-Id: Ie9cad48822055f8d06b45301e5486ff032f3b5fb

GitOrigin-RevId: 8ee9953d0bb327730b9f51ea6a5ee887297d57c6
(cherry picked from commit 8d026313f0301e5d0986d1460e8f25c393b88e58)

Space-RevId: f18829ecef3bf1febee5d449249d8d2ac7d911c3