Fix ResourceImportDialogViewModel validation message
The validation message in `ResourceImportDialogViewModel` was not being
updated correctly when the asset set was empty. This was because the
`doValidateAll()` method, which has side-effects of updating the error
message, was not being called when the asset set was empty.

This change ensures that `doValidateAll()` is always called, even when
the asset set is empty, so that the error message is always up-to-date.

Additionally, the format of the validation message has been updated to
be more consistent.

Fixes: 442763194
Test: ResourceImportDialogViewModelTest
Change-Id: I0f782b8d63daa74fbd4d6f441f15cb012720aca5

GitOrigin-RevId: 293b33be396526a494d4217a08f3689704f993ed
(cherry picked from commit e8c806369d3ae32a00f2e32c4db9d135f04c8a21)

Space-RevId: dcf34216f7937206856568eb7cad5902a077b0af