Add files required by GenerateComposePreviewAgentEval to
... SimpleComposeApplication.

The Conversation composable has parameters but no preview. This will be
used by the eval to check that the model will properly understand how to
pass sample data for the parameters. Furthermore, the CL adds the
definition of the parameter in a separate file to ensure the model is
capable of getting that definition.

Fixes: b/433947229
Test: GenerateComposePreviewAgentEval
Change-Id: Idf321c971d9eb1530a033bf0cb8b3b7211b01777

GitOrigin-RevId: 6798a29e1cd7cec8ae55c45baf7f843038a5805b
(cherry picked from commit 567ee6b883cd45f10ee9d007488c1acfea97af28)

Space-RevId: f1e8c6c5f517dcf7da6f1e964a32c1d181064745