Fix compose previews order to be case-insensitive and locale-aware
The previous lexicographical sorting of Compose Preview elements by
their display name was case-sensitive and did not consider
locale-specific sorting rules. This resulted in incorrect ordering for
previews with names containing mixed-case characters or accents.

This CL fixes it using a `Collator` to ensure the preview names are
ordered correctly according to the conventions of the specified locale.
The locale is determined from the preview's configuration, defaulting to
the system's locale if multiple different locales are found or not
specified.

Bug: 438684987
Test: PreviewElementSortingTest
Change-Id: If9a527d1cbc826eb152116f1178d445ca323b085

GitOrigin-RevId: ef1e395da654dcac4d4052a8fece37dc151f4222
(cherry picked from commit fe09a7ff3b521c46f8e400815c68f441e093ec1a)

Space-RevId: 24969264a695c5f312b701bf46c0de67525b74d8