Fix slow operation in NavEnumSupportProvider
There are three places where EnumSupport.values is called.
In two of them, it is called from the background thread.
In the third case, it's not called from the background
thread, but we it uses IdEnumSupport directly and does
not use ClassEnumSupport.

Fixes: 321695920
Test: NavEnumSupportProviderTest
Test: ComboBoxPropertyEditorModelTest
Test: PsiDropDownModelTest
Change-Id: I9b58d624b33f1edf0448f44e952928e3547c16bc

GitOrigin-RevId: d13862f29fefac580bb46fb8769bb3aed76dea47
(cherry picked from commit 0148f5711c0348274e4fa4f8f0a48051962b9ae8)

Space-RevId: a509c3e627bb2cc15da190f08390765824b67e54