Update preview status messaging
Some states like out-of-date and needs-build are detected relying on
not granular aspects, like file modification tracking or last build
status. This means that in some cases we may not be certain about
whether the preview is out of date or not, or whether a new build
is needed or not to display the previews correctly. What we actually
detect is the possibility of this being the case.

In this context, the tooltip messages of the preview states are being
slightly modified to better reflect the context in which each state
happens.

Bug: 434139799
Test: CommonIssueNotificationActionTest
Change-Id: Ic344d0d2b59383c7cac603c3f0400f583a9f3fde

GitOrigin-RevId: 95be9390826b8cb2bde71689ff1a46ba24790d68
(cherry picked from commit e655a716b0736e3ca89bf3caec5b91fe192dd24e)

Space-RevId: 9bd2bf92eb7d686b69dca217e91d06f51a19388b