Remove the toValues lambda from findAnnotatedMethodsValues
... and rename the method to findAnnotatedMethods.

The toValues calculation is now the responsibility of the caller.

The toValues parameter was preventing caching from working in cases
where the toValues was a lambda. This is because the lambda is different
on each call, meaning the key was different on each call too. This was
not caught by the tests, as the tests were passing functions which have
the same identity on each call.

The toValues was computed within findAnnotatedMethodsValues under a read
action. This CL adds readActions where needed as the toValues has been
moved out of  findAnnotatedMethodsValues.

Bug: 381827960
Bug: 397205975
Test: existing
Change-Id: Ifcf32ebe25dc268e5ddef8c6da2fbe74d90e6988

GitOrigin-RevId: 75c653d41fb12c321e152b2ee83585a9a032c9eb
(cherry picked from commit 9441ef8432f4a4948fa7a9f84accdaf1577fd901)

Space-RevId: 431e6e1c2f793b939896fa2fa5a3eb6fb04ae369