Fix Composition error reporting in interactive
Composition errors are "swallowed" by the hot-reload mechanism to allow
compositions to keep working. This change uses that mechanism to
intercept the errors and report them back to the user. It also clears
the errors to the composition can keep progressing.

This also updates the ErrorQueue mechanism since it was broken for
interactive renders. Because of the MergingUpdateQueue, the update
errors call would happen too frequently preventing the merging queue
from dispatching the errors to the UI.
The change updates the timing of the updates in the following way:
- The first update will happen immediately. Before the change, the first
  update would have taken 200ms as this was the merging queue delay
  time.
- After the first update, the queue will wait 1s before the next update
  happens. This avoids that quick updates change the errors in quick
  way. Before this change, no updates would happen until after 200ms of
  the last update.
- If the next update happens after 1s, it will happen immediately again
  and will not wait 1s.

Fixes: 249894454
Test: Existing and manually verified
Change-Id: I3f97cb445dd7d8d58f0835f2762522359caaea26

GitOrigin-RevId: 0b49b1e3aa3d983fb72df77b08d9cf6968eef7d4
(cherry picked from commit 5db363cca23bf3e9de8cd894ffc51460c53c3864)

Space-RevId: 99ae7e537d71426118dcebcac42d9d4eae0fb3f6