[AUA] Reduce difference between tests and production
Don't disable forced upgrades generically in all tests, because that
makes it not possible to actually test the behavior of components of
the forced upgrade.  Instead, mock the method that displays the
dialog where necessary (because it triggers an IllegalStateException
because of headless dialogs not being considered as modal) but
otherwise allow the rest of the flow to run.

This means that one test now picks up that there will be quickfixes
for upgrading AGP.  (This is probably a temporary state of affairs as
the next CL changes how the forced upgrade state actually runs.)

Bug: none filed
Test: existing tests adjusted
Change-Id: I552687dd65e47625e78982cda773820f0e56fef2

GitOrigin-RevId: d360a45124a5cb4c0a43aba9a420e1a3f5c6db6c
(cherry picked from commit 21185e25fe699a7191ccb15264d588c339bd90a3)

Space-RevId: 602b7726295418f503edf267dbe2dab7fc3c545a