Fix flaky test on windows
The flakiness was caused by the ProjectSystemService triggering a root
change while a dialog was showing. This caused some issues due to a
write-unsafe context as a write action was being performed in a
different modality than the dialog.

This CL fixes the issue by ensuring the write action is performed in the
non-modal modality, to ensure that it waits for the dialog to close
first.

The root change is already invoked later, so waiting a bit longer should
hopefully not be an issue.

Fixes: b/447592104
Test: tools\base\bazel\bazel.cmd test \
 //tools/vendor/google/ml/aiplugin/android\
 :aiplugin.android_tests__other\
  --config=remote --runs_per_test=1000
Change-Id: I14ebc36cf659ff76b045959c28672d89003fff68

GitOrigin-RevId: f6f4186aad97a9017b05a279581a36d2a1fe871a
(cherry picked from commit bb60b7fb342584f32c6418f792ca642699b4d862)

Space-RevId: 152dbe1a7c2a8cfc57649765637d103c5f0e1d3d