Fix BuildProjectBleakTest
There's a new field in ProjectNotificationsModel which holds
notifications; make sure we ignore references from it.

Also, the class is ProjectNotificationsModel, and not
ProjectNotificationModel.

Finally, since these notifications in any case hold references to the
project, make sure that we expire old notifications when we are about
to create new ones so that we have a constant number of references.
This isn't strictly speaking a leak but pacifying the leak checker is
probably easier than informing the leak checker about the semantics of
project notifications.

Bug: 445656379
Test: this is it
Change-Id: Ib6833d88fc5e4abb00bc489aafc9d569f425320a

GitOrigin-RevId: e7761e73b88a9eb0f770e8e35b4a6722fbda7003
(cherry picked from commit a2b532eedbec53cfd26549969a031a83f3114b2b)

Space-RevId: 87e74458e0292a0f48093d69722129fba56be621