Separate Read And Write Actions
EditorFoldingDetector.detectFoldings() can take a significant amount of
time if the line span is large.

It's currently running in the EDT as a write action but it can be split
into a read and write actions where the read action is run in the
background.

Fixes: 449159026
Test: Existing
Change-Id: I21f3c1ee82b0d56073b32f2e14bf733fc945dbdc

GitOrigin-RevId: 45a17ea15ff37dbfdeafa7fba5901ad0e007c8b8
(cherry picked from commit 3abf2424fcb62053b16658f2248e41e35068dc6f)

Space-RevId: 3735f18f85a23bdc3beb2289689242ad7c5fc594