Decouple resolved dependency storage in workspace model
This will allow separately updating the dependencies without incurring a
performance hit.

Implemented by always storing the "core" GradleAndroidModel, without the
dependencies, and store the resolved dependencies separately. The actual
implementation is then mapped using ExternalMappingKey's and then
accessed by the GradleModuleSystem.

This change comes with some seemingly unrelated changes, but they are
all required (and there were no clean way of including them in
a separate change)
- Make GradleAndroidModelImpl compatible with the new version of the
code generator (it was updated with the new platform to be compatible
with K2 and has slightly different behaviour) and regenerate entities
  - Make AndroidModelFeatures final
  - Make GradleAndroidModelImpl a data class
  - Replace @Child on child entity with @Parent on parent entity
- Update sync contributors to partially create GradleAndroidModel first
and then attach the dependencies to WSM
- Fix a bug in AndroidSourceRootSyncContributor that meant we missed
updating GradleAndroidModel/GradleModuleModel when it changed
- Replace equality check in GradleAndroidDependencyModel to compare with
the core model instead. It's now a data class so we can do it.
- Adjust GradleAndroidModel factory signatures. This will be further
changed.
- Make patchForKapt helper method visible for a single variant too.
- Removed the equality check in IdeModelDumper because the new hierarchy
doesn't allow it anymore. This check will not be actually required
anymore with further cleanups as well.

Bug: 384022658
Test: Covered by existing integration tests
Change-Id: Ib11a8a1514f0f138cba580cb0a7efe9fc4a30cb4

GitOrigin-RevId: 8fac75798b6c01dc6728ed76230d96606c54e17b

Space-RevId: 4d88c71fdadd1c39819fe6b82ea345a6c2c76210
(cherry picked from commit 3ca7de7e79b99cbbd613f7959b6ffacf8ec996d1)