FR: Add top border to Memory Profiler tabs
Adds `AdtUiUtils.DEFAULT_TOP_BORDER` to the `JBTabbedPane` within the
`CapturePanelUi`. This provides a visual separator between the "Table"
and "Visualization" tabs and the Memory Graph. This improves the clarity
and ease of identifying the tab area, especially when the panel is
resized, addressing the user feedback in b/450771004.

Fixes: 450771004
Test: Manually verified.
Change-Id: I434fc61ab950563d95c4338d1e56e2cd0d7f2ba0

GitOrigin-RevId: 237f588ca50df993f782d24faeb73e3c17075b84
(cherry picked from commit 4e6e648fabc7840861a3a4d65ace456c67873e56)

Space-RevId: c51ad052f591fae1718a6fff6ee8d39442fa02a5