Move the zoom masks logic into ZoomController
Lot's of bugs where related with the race condition between
DesignSurface and the zoom-to-fit that has being notified too early
before the DesignSurface had the right size.

We adopted this logic only for Preview but we are now extending to all
the NlDesignSurfaces.
This will also fix issues with surfaces showing big zoom on drawables

Fixes: 445109082
Fixes: 374663154
Fixes: 425997179
Test: DesignSurfaceSettingsTest, VisualizationFormTest, ComposePreviewRepresentationGradleTest
Change-Id: I2805bb93775c3a9597ce53be4e5a411108a7fe09

GitOrigin-RevId: 2539f806fb666e1c2023c5c2b150412fe88f1e9d

Space-RevId: fb00bea4a0276bf1ba04ee0cc8969ebde57edab0
(cherry picked from commit 515bc5853858a6b583578c6314388186802dcfbd)