Ignore template filter for the Journeys template
There is logic in 'NewAndroidComponentAction' to filter out any
templates that are produced from ScopeType.MAIN source providers.
This means that when a Journey file is selected, there are no valid
templates and an exception is thrown.

We can fix this by bypassing the filter for the Journeys template.

Fixes: 451538323
Test: Added
Change-Id: I4acbd3a25ecbf8de2180b17428cd62a10bdc9521

GitOrigin-RevId: 17766f0f0d2cacf1f84b9a549a2e89fe5d7fd359
(cherry picked from commit ca4b215dad89150c9a6f2548173cacb6a0be69b3)

Space-RevId: 043c5a8cf440df6f592f7c78f95ab80104e03d49