Call action.update on actions used in the empty panel
... and only use them if they are enabled and visible.

In ComposePreviewViewImpl we had an implementation that
just checked whether there were composables in the file
assuming that there were no previews. That assumption
is not entirely correct, as we can end up in the empty
panel when there are previews (when there are syntax
errors).

Instead of trying to duplicate the logic across the
action and ComposePreviewViewImpl, this CL attempts to
reuse the action logic to check if it should be visible
and enabled.

Fixes: 450783824
Test: Added regression test to ComposePreviewViewImplTest
Change-Id: I90a992a69a213faf15d879362938e902d8cc1a4a

GitOrigin-RevId: b440af605fa7f523a4ed849be8d8e31b0dabbcef
(cherry picked from commit af2061bdf60f76f82ea9c0cce0c977fb88a32dcf)

Space-RevId: b7daacadef8386eb7ecc89687be3fdac575e4d53