Use smartReadAction when checking if
... ComposePreviewRepresentationProvider can create a reprsentation.
It's not enough to check for DumbService.isDumb(), because the project
can return to DumbMode between this check and when we actually rely on
the indexes, causing an IndexNotReadyException to be thrown. In addition
to the DumbService.isDumb() early return, we should wrap the read action
into a smartReadAction instead. This way, we will suspend until it's
possible to obtain the read lock in smart mode and only then we'll run
the action.

Fixes: 451626256
Test: N/A
Change-Id: I370c1a3311dbe545badeb32e4dd904480f431344

GitOrigin-RevId: 93bd2cde9d4ed9af2ec7d1a0d1f212d268f58bd5
(cherry picked from commit 390e6c73d1897c4b9cc1db6d8cfb18e87fc959aa)

Space-RevId: 25ae13a93274c47fe61e37f9f7bf57c15fc9eb54