Show warning in Wi-Fi Device Manager entries if device not up-to-date
adbd with the wifi v2 changes publish an mdns service with "2.0" in the
mdnsServiceVersion.

Old devices may not be eligible for an update, we can't detect that so
the warning message won't be actionable for such devices. It should be
ok to show the message since it doesn't stop the pairing process.

Bug: 412571872
Test: WifiPairableDeviceProvisionerPluginTest
Change-Id: I7fea2929405df39624fb0bb3d2110246c83e3928

GitOrigin-RevId: 014ac1c415e40e256cdd737df63bf33cd6364035
(cherry picked from commit 5f161e719cd86c9e88a136ce7c25e6d89c6867c2)

Space-RevId: 894080513c25080e008a68815d72ca1b95568edd