Show warning message in pairing dialog if device not up-to-date
If mdnsServiceVersion isn't 2.0, we show a warning to update the device.

Old devices may not be eligible for an update, we can't detect that so
the warning message won't be actionable for such devices. It should be
ok to show the message since it doesn't stop the pairing process.

Test: WifiPairableDeviceProvisionerPluginTest
Bug: 412571872
Change-Id: I21e12ce10909398d9a110bb2cf53c2125b7cb56a

GitOrigin-RevId: 7d42a4e3b4ed3b9ef345bea3a114026227ff4b39

Space-RevId: 1b6549534ac5992d13e641ade515f8c56571710d
(cherry picked from commit 2ae36f37006b623713fdade6d6c2066ee50230f0)