Show warning in available devices dialog if device not up-to-date
If mdnsServiceVersion isn't 2.0, we show a warning to update the device.

Old devices may not be eligible for an update, we can't detect that so
the warning message won't be actionable for such devices. It should be
ok to show the message since it doesn't stop the pairing process.

Bug: 412571872
Test: WifiAvailableDevicesDialogTest
Change-Id: Ie46945843f614820e8e40dc865b45e2b9beb29cb

GitOrigin-RevId: a32c3da48cb9256972348439c6e3a163035bccdc

Space-RevId: ff8e342b0afa2a8d0387534775b1a31a5d96d24c
(cherry picked from commit 4f56ea8cd8fac9ffce327d651e29a7552ff06fdf)