Remove parsing of build files when detecting FTL presence
Retrieve the information from models through the project system instead.

This is an expensive operation and close to the critical path of sync
and needs to be removed for phased sync and is an improvement regardless
of it.

Bug: 384022658
Test: Replaced some unit tests with integration tests
Change-Id: Ibcdaf2eebd18f5697569c329704567447b44a77f

GitOrigin-RevId: 22a31d7ad809ffeaaf7f4c7e1b684eeef2edfbab

Space-RevId: f76e62d213db1827187c468bc676921bfa1cb339
(cherry picked from commit bfe543039453de2ff69353f2ed28c01188a687e2)