Ignore right clicks when cursor is not on panel
When we have multiple displays and on-device rendering turned on
there is going to be a renderer panel for each display. When the
device sends the right click event, it will be received by all the
panels. But we only want to show the popup in the display where
the user actually performed the right click.

This change ignores the right click for panels that don't have
the cursor ontop.

Test: OnDeviceRendererPanelTest
Bug: 443263320
Change-Id: I3a2990e632f173e751860d78076f8802e5452ff1

GitOrigin-RevId: f70e5338784dfc2739dbec9f828fec8a1e3b05fd
(cherry picked from commit 1823b779e180964097739f8cfe4fbae0059a8422)

Space-RevId: 801990831d92953b7ff6b24419e7e40fc093a355