Fix/polish collapsible code blocks
This CL fixes the collapsible code blocks, which show the scrim in the
incorrect state. It also reworks how the collapsing blocks are handled
completely, to avoid subtle issues with the content, such as the scroll
bars being visible at the bottom when the content is collapsed and they
should be well below the fold, and thus invisible.

Instead of simply setting a max height modifier on the content, it uses
a custom Layout that always measures and places the content as if it
were not collapsed, but still caps the height of the block in reality.
It also makes the content not focusable when in the collapsed state, to
avoid the focus and screen readers being caught by invisible components
that are in the collapsed area. It's not possible to keep visible
components focusable, so it has to be an all or nothing approach.

Further, this polishes the existing code to:
 * Remove hardcoded UI text where possible in favour of messages bundle
 * Remove hardcoded test tags in favour of the usual LayoutTags pattern
 * Swaps out the custom, cobbled together implementation of the bottom
   "expand code" button with a tweaked ActionButton, which has better
   a11y properties and actually adheres to the LaF by not using the base
   generic Compose indicator for its states (hovered, pressed)
 * Fix clipping issues with the bottom button, that expanded over the
   block's rounded corner
 * Cleaned up the code a bit in general

After: http://go/scrcast/NTIzNDc3MzI3MzM0NjA0OHwxZTMzZmYyNC0xOA

Bug: b/454008417
Test: maintained existing
Change-Id: Icbf3d5876c9e0a02417edb909cdabcae8a6c74db

GitOrigin-RevId: 8e5b8671273a454ae53bfa64a97afa066f7a6a97

Space-RevId: ce6c1b167687985bd49bc9b77f0d97401f0e5b88
(cherry picked from commit e6eeb5033fad5e7588e547fda10749dba7d510ad)