AndroidPlatformSdk without project view set
Make it possible to obtains AndroidSdkPlatform from an skd string and
an minSdk value without providing a full project view set.

The implementation may still internally request the location of the
top level project view file to report it to the user. This is a leaky
abstraction that allows breaking the dependency on the query sync
project on the project view set, which makes it difficult to find out
which values from the project view set are consumed by the query sync.

Bug: n/a
Test: existing
Change-Id: Ib2103284623a379e5818803a6be4b34be916bed7

GitOrigin-RevId: 022664031f8eb8fb0e54851384f022e24df1c399

Space-RevId: fd08b5baed65c3299d6cbc29dcce376718f4ee2f
(cherry picked from commit 34dbf4d9aed1a415996ba9ab6d4581b3a7cdd9a3)