Change AI Actions dropdown to default group
Updates the ComposeStudioBotActionFactory to allow AI Actions to be
rendered as a flat list, in addition to the existing dropdown display.

This is necessary for integrating the actions into the preview overflow
menu where a flatten list of AI actions is required.

Bug: 455568136
Test: PreviewSurfaceActionManagerTest
Change-Id: I5d5db1702a40e18eea69dc411672e4fcba969ed8

GitOrigin-RevId: 1fbe55516ab4627381401cd08462f7a5b19e7642

Space-RevId: 40310ed37569e205c38dbff19ccc777df5fdb580
(cherry picked from commit ff256ab316aac01a1f34fea1313cc04c4644da26)