Fix regression introduced by ag/36611858
In AndroidSdkFromProjectView.java the check for projectViewSet == null
was removed and replaced with call to getProjectViewForUserMessagesOnly.
When project is null the first line of that method fails with the call
to ProjectViewManager.getInstance(project). This fixes that.

Bug: N/A
Test: Import new Google3 project
Change-Id: Ic3740095cd98c67390af63c1b59613857b8b91e0

GitOrigin-RevId: 4206d86864297c11f71ac1b683f6aa87db40a397

Space-RevId: f7e7fde482e79d0372895ae8d8d68a49ed29d493
(cherry picked from commit df6af1749e20a1a2e744a84086920252cc38d653)