Add E2E integration test for LeakCanary task
This change introduces a new end-to-end integration test, LeakCanaryTaskTest,
for the LeakCanary task in the profiler. This test validates the entire user
workflow within the IDE, ensuring the feature's stability and preventing
future regressions. The test automates the process of starting and stopping
the LeakCanary task on a debuggable application (minApp) and verifies
the correctness of each step.

Bug: 454195428
Test: New test LeakCanaryTaskTest passes
Change-Id: Iaa6fe78f332b84cbfdb426f4d5244ff984645990

GitOrigin-RevId: 8015c2510eee026005c4346b36ded093e0cb3d51

Space-RevId: 1af9e87b7196b2bd6adb43cb2d1208fbeccbc069
(cherry picked from commit 849eb88d4b699e85727b2a7ea3d4454cefe7586e)