Disable Android template actions when the context has no module
If you close all files in the editor and then search for a template
using the 'Search everywhere' feature (double tapping Shift), then the
action for the template will appear, but will do nothing when clicked.
Unfortunately, the action will still appear in results when setting
`isVisible` and `isEnabled` to `false`, though it will appear greyed
out and indicate that it is disabled - which is an improvement on the
current situation.

Fixes: 435599346
Test: Added
Change-Id: Ibe547327687770000595b40a5989b442c0e8b9b3

GitOrigin-RevId: b5fd318e40651e1c399b7a9120f804a60b7700d3

Space-RevId: f90adf8162b0cbf09749c430fdbd423d2ddc3d13
(cherry picked from commit e8647c03546cbe40f205e7b2b734ff2f5526a3c6)