Remove unnecessary check
The waitForCondition will fail the test if the command is never
received. It is not necessary to check that the command was the
last command received (sometimes another command is sent before
the check is made).

Fixes: 456565960
Test: OnDeviceRendererModelTest
Change-Id: I385b57bb1a5bfb60761ed3625be8d5ca96fa563e

GitOrigin-RevId: b82f64bddd84d5b981c6f663c174f8d83d417f16

Space-RevId: 6d5e5d0d458871eea158a13bde02b08b147a4b89
(cherry picked from commit 39cc4f610a3823ac86e101264b9e43c7cad31b52)