[kotlin] cherry-pick changes from AS about detection new KMP Android Target
This is necessary to verify IDEA integration with new KMP plugin faster.

IGNORE on Merge Conflict!
This is a copy-paste of the code from AS, it should match. but when not
drop this commit in favor of the one from AS.

^KTIJ-36188


(cherry picked from commit 39ff520d510a87dfd648380e02f85acc1901bba3)

IJ-MR-180641

Space-RevId: 14fad96810feb10a438c478e1e9a5c0c6645f4b3

GitOrigin-RevId: 0767a4f876ec6d0fb2a0ac5c8816eca22d9694cc