[kotlin] cherry-pick changes from AS about detection new KMP Android Target
This is necessary to verify IDEA integration with new KMP plugin faster.

IGNORE on Merge Conflict!
This is a copy-paste of the code from AS, it should match. but when not
drop this commit in favor of the one from AS.

^KTIJ-36188


(cherry picked from commit c30a0d43f217a0cace454879429aab4d800270f0)

IJ-MR-180641

GitOrigin-RevId: 14fad96810feb10a438c478e1e9a5c0c6645f4b3