Fix flaky test: StateInspectionPanelIntegrationTest
This test revealed a problem with the model:

When a new request is emitted via stateReadRequested the cache
could react and send a (cached) stateReads result before the model
has a chance to handle the new stateReadRequested.

If the model handles the stateRead result before the stateReadRequest
the model may overwrite the result with a waiting state. Detect this
in the model.

Fixes: 458142095
Test: StateInspectionPanelIntegrationTest
Change-Id: I387ec6e03d63e8980eec5eb58e254ca0f3e9d67f

GitOrigin-RevId: 5142cf97d4d80f8eb8d88ce740bd531256e47967

Space-RevId: 9b85dfb9c28aaa5a22dd8e55b837600a3484ea7c
(cherry picked from commit 0d63d1fa50265800bc7edc0243e585f8294023cf)