Collapsible floating toolbars
When the "Make zoom and XR toolbars collapsible" flag is enabled, the floating toolbars by default show a single button and expand horizontally when the user clicks on it. The toolbars stay expanded until the user click of the ">" button on the right side of the toolbar.

Bug: N/A
Test: included
Change-Id: I9ea855f400d8a9c4a3cf35a8ae37b0d01f8c6bac

GitOrigin-RevId: c7d729e590538378465a996a2215ee9db9a4214c

Space-RevId: 4681b05b3b215e58a7cc2dc41bbd92daf041b9ad
(cherry picked from commit bf400aeb4e46b4096d57bda2505057bc5460d26e)