Do not run EditorBasedTableCellRendererTest on EDT
The function:
AbstractExpandableItemsHandler.updateAlarm.waitForAllExecuted()
was changed in 2025.3 to join on the updateJob where it before
just waited for the updateJob property to be null.

That means we can no longer call this function on the EDT since
that would block the job we are waiting for.

Fixes: 460420821
Test: EditorBasedTableCellRendererTest
Change-Id: I3c442602babb80ae56ffdc1662c5b7bfb193df4f

GitOrigin-RevId: 511e91b37fe15e0772fd79d32239fac59dd71222

Space-RevId: a28a195e29b101e8bf8800184b0f3d99f3383188
(cherry picked from commit 93ccea280fcb5f007618e387c847283530b744e4)