Defer to platform for isAllowed() in analytics
Background: we have a platform patch that intercepts calls to
UsageStatisticsPersistenceComponent.isAllowed() and redirects them to
AndroidStudioAnalyticsImpl.isAllowed(). However, this platform patch
is unnecessary because the platform consents and the Android plugin
consents are already kept in sync at all times. The syncing happens in
AndroidStudioAnalyticsImpl.initialize*() at app startup and continues
in AndroidStudioAnalyticsImpl.update*() whenever the platform consents
are modified by the user in the Data Sharing settings dialog. To be
extra sure, I've added a new unit test which passes both before and
after this change.

Side note: the platform consents are treated as the source of truth
(ever since ag/4768793). One could wonder whether the syncing should
happen in the opposite direction eventually, but that is outside the
scope of this change.

Bug: 460518368
Test: AnalyticsSettingsUiTest.testConsistencyWithPlatformConsents()
Change-Id: I1ae19c12c5fc9d67aca4be83cf7a3652cbca738c

GitOrigin-RevId: 42b6dcd841d0d9af4b4f6a79d1f0bca77e1463d3

Space-RevId: ea90110008a55c9aec95e8ff232ca7eb31db70d1
(cherry picked from commit 57620a286d40dd7c35569fea48e03561de1fc378)