Hide an action in Gradle projects
The AssembleIdeaModuleAction almost always hides itself in any case;
it would only be active when the project system is a
GradleProjectSystem, but without a requirement for Android models (so,
a pure Java Gradle project?).  In those circumstances, the
AssembleGradleModuleAction is active, so there seems to be no benefit
to this action.

The 10-year-old comment about not hiding this action because of an NPE
appears to be obsolete.

Bug: 452574547
Test: existing
Change-Id: I880d5dd14064a670b8db71a86aafbc869ca829e7

GitOrigin-RevId: 3fd4f7a69931d0e8ff38100147612728a4a5e07c

Space-RevId: 46c5c51f464cc02642b4d5983a8fd4d787dce159
(cherry picked from commit 377eb758d5de6fbf9e487c13bbd2ba3e5b8cbc7b)