Do sync for maven fix outside of write action
When undoing/redoing the Maven import fix, we need to make sure the
Gradle sync operation happens outside of the context of the write
action. Since the undo/redo doesn't need to wait for the sync to
complete and just needs to trigger it, pushing it to a subsequent EDT
invocation works.

This was caught by test failures in the 2025.3 merge.

Fixes: 461786379
Test: AndroidMavenImportFixTest
Change-Id: I1cc3e0f711e3d37d9aec8973c2dc5e9052557315

GitOrigin-RevId: 5e1e43fa6123b4e53e4dda8c33245970460e1044

Space-RevId: 4dd8c553f235354e003cf8dd479510635d30463c
(cherry picked from commit 917a4007f9582596af6398c752229998cb5357a8)