Add LeakCanary presence check to profiler task
When the LeakCanary profiler task starts, check if the LeakCanary
library is actually present in the application's dependencies.

If the library is not detected, a message is now displayed in the
details panel instructing the user to add the required
`debugImplementation` dependency to their build.gradle file. This
prevents user confusion when the task starts but no data appears.

The LeakCanaryModel now sends a CHECK_LEAKCANARY_PRESENT
command on task start and listens for the response event. The UI layer
(LeakCanaryScreen and LeakDetailsPanel) was updated to observe this
new state and render the appropriate message.

Test: Manually tested by running the LeakCanary task on an app with
and without the dependency. Verified that the correct UI message
appears when the dependency is missing and that the task functions
normally when it is present.

Bug: 461789645
Change-Id: Ia69fa63e809f98e7cc433c49c89f4a484179b304

GitOrigin-RevId: bc717479e9d099b13ec261c2ecb21125cf03e2d2

Space-RevId: 01e9a575eb848c240da38546846fba0925e513f3
(cherry picked from commit c1d04137608934f91875f996cb437a3ce6a6fb32)