Fix quick fix tests for 2025.3 merge
The RenameTo fix is no longer a fix that needs to be unwrapped in
testing due to refactoring in the 2025.3 merge.

This change allows the test to pass both before and after the merge, and
I'll follow up later cleaning up the old version of the code.

Bug: 461785883
Bug: 463392037
Test: AndroidManifestDomTest.testSpellchecker2
Test: AndroidLayoutDomTest.testSpellcheckerQuickfix
Test: AndroidValueResourcesTest.spellchecker2
Change-Id: Id2033b596411ffe728dafccf8daabb133a10bff8

GitOrigin-RevId: c4c236e7c14be14145b2cd474126735dfbbef996

Space-RevId: ddcc906d6e12e91586e68b7dd886c9054be7887c
(cherry picked from commit 1040bcc0b943b551fcb6a7d3b7d9ac79b3003e68)