Work around missing information from the KotlinModelBuilder
We do not pass the needed srcSet information in the kotlincompilation
data to the
KotlinGradleModelBuilder in order to setup visibility onto other sourceSets.
Because of this, and in order to be able to add support for having
visibility onto "friends" sourceSets, we will be setting this
information at the level of the kotlin facet configuration (in the data
service).

Bug: 357930912
Test: reflected on the snapshot tests
Change-Id: I715fec777475c1835f0c4352f394da0c1c8c8c5e

GitOrigin-RevId: 9b3d8b628a6339ca3c2c754eafd65547fe33044e

Space-RevId: 9b3c1840253a91a9743085a0d2b37d3932e087bc
(cherry picked from commit e9e0d9660e253a9b7e0d6ea2c9856bf3173ddd8f)