Hide AI actions when opening a Compose Panel on test files
The "Generate code from Screenshot" and other UI Tools Agent actions
were incorrectly displayed in the empty state of the Compose Preview
when viewing a test file. These actions are not applicable to test
classes and should not be shown.

This change prevents these actions from appearing by introducing a
check to determine if the current file is a test file (isTestFile).
If it is, the AI-powered actions are hidden from the empty state view,
aligning the UI with the intended functionality.

Fixes: 452871383
Test: ComposePreviewViewImplTest
Change-Id: If7176bb8fd5b9efc252371bb77f8adc57aa86274

GitOrigin-RevId: 1ce56f3f933af5e8e07509882d762c8fa915b2a9

Space-RevId: f67ea6e30d435bf67eb696bb5465fca5bd187969
(cherry picked from commit 130f1da3f995f90ead2420011b5caba8f313b0ae)