Resolve projects variants by priority order
Take into consideration the incoming dependencies total count (provided
by the DeclaredDependencies model) to propagate the variant
request across the dependencies.
We do this by implementing a topological sorting algorithm in order to
insure that dependencies are only resolved once all of their consumers
have been resolved. This insures that conflicts are avoided, and
priorities are correctly handled.
At this stage, we are just falling back to the default variant when we
do not have the match for the initial requests. In the next steps we
will be using the matchingFallbacks, then the missingDimensionStrategy
to attempt resolving conflicts, before falling back to (when possible)
the default variant.

Bug: 467047467
Test: N/A
Change-Id: Ib638d19cee2d98f10269a86a6b06dde7a1e813ae

GitOrigin-RevId: a38cc32c6c965649ac1fbfee305964233fa8ecdc
(cherry picked from commit 36f227b9548f56344244247d326f77994fe41038)