Use MatchingFallbacks in variant resolution
When there is no direct match for the BuildTypes or the productFlavors,
this will be useful to use the matchingFallbacks (ife declared) to
fallback into other variants when there is no direct match.

this is same as what we were doing just before: this change now also
caches (for each selected variant) the BuildtYpes(s) / productFlavor (s) with their
respective matchingFallbacks. We also store, for each matchingFallback,
its occurences across all the projects that have declared it, and this
is done in order to prioritise the fallbacks that will result in the
maximum amount of projects' variants resolved.

Bug: 467047467
Test: N/A
Change-Id: I481251c2c8105a41a037461165b8ccb7a5aa18a8

GitOrigin-RevId: a0a763448d72bb04c578989a49d9103cab5ca6a6
(cherry picked from commit 992bc53b2698d76828bbe93b992e27baa099dcf5)